FIX/mcp4728: A sda pin stretching workaround
After a valid read_address sequence the MCP4728 wond release the sda pin to high. To bypass this problemm we add some extra stops until DAC releases the pin. This is not the best of solutions but it seams to working. We also have add the same procedure in the begining of _init() in order to be sure that the MCP starts with sda released pin.